Fix: Consolidate config.py and config_v2.py into single configuration module
Resolves incomplete migration from PR #84 where config_v2.py was created
but never fully consolidated. This caused confusion about which config
module to use and risked validation logic drifting apart.

Changes:
- Merged all unique models from config_v2.py into config.py:
  - ProfileMetadata
  - InsuranceLayerConfig, InsuranceConfig
  - LossDistributionConfig
  - ModuleConfig, PresetConfig
  - WorkingCapitalRatiosConfig, ExpenseRatioConfig
  - DepreciationConfig, ExcelReportConfig
  - IndustryConfig and subclasses (ManufacturingConfig, ServiceConfig, RetailConfig)
  - ConfigV2, PresetLibrary
- Updated all imports from 'config_v2' to 'config' across:
  - 22 Python modules
  - 6 Jupyter notebooks
  - 4 documentation files
  - README and tutorials
- Deleted ergodic_insurance/config_v2.py
- All tests passing (105 tests total)

Impact:
- Eliminates configuration confusion
- Single source of truth for all config models
- No breaking changes (all imports updated)
- Prevents validation logic drift

Closes #227
