+++ ergodic_insurance/tests/test_config_validation.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from ergodic_insurance.config_v2 import (
+from ergodic_insurance.config import (
     IndustryConfig,
     ManufacturingConfig,
     RetailConfig,
