+++ ergodic_insurance/tests/test_depreciation_tracking.py
@@ -2,8 +2,7 @@
 
 import pytest
 
-from ergodic_insurance.config import ManufacturerConfig
-from ergodic_insurance.config_v2 import DepreciationConfig
+from ergodic_insurance.config import DepreciationConfig, ManufacturerConfig
 from ergodic_insurance.manufacturer import WidgetManufacturer
