+++ ergodic_insurance/tests/test_tax_handling.py
@@ -7,7 +7,7 @@ tax-deductible business expenses.
 
 import pytest
 
-from ergodic_insurance.config_v2 import ManufacturerConfig
+from ergodic_insurance.config import ManufacturerConfig
 from ergodic_insurance.manufacturer import WidgetManufacturer
